requirements.txt
fastapi
uvicorn
pandas
boto3
python-dotenv
requests